feat: add command completers logic
