+++ xonsh/completers/init.py
@@ -1,12 +1,10 @@
 """Constructor for xonsh completer objects."""
 import collections
 
-from xonsh.completers.pip import complete_pip
 from xonsh.completers.man import complete_from_man
 from xonsh.completers.bash import complete_from_bash
 from xonsh.completers.base import complete_base
 from xonsh.completers.path import complete_path
-from xonsh.completers.dirs import complete_cd, complete_rmdir
 from xonsh.completers.python import (
     complete_python,
 )
@@ -15,6 +13,7 @@ from xonsh.completers.commands import (
     complete_skipper,
     complete_end_proc_tokens,
     complete_end_proc_keywords,
+    complete_xompletions,
 )
 from xonsh.completers._aliases import complete_aliases
 from xonsh.completers.environment import complete_environment_vars
@@ -32,9 +31,7 @@ def default_completers():
             ("base", complete_base),
             ("skip", complete_skipper),
             ("alias", complete_aliases),
-            ("pip", complete_pip),
-            ("cd", complete_cd),
-            ("rmdir", complete_rmdir),
+            ("xompleter", complete_xompletions),
             ("import", complete_import),
             ("bash", complete_from_bash),
             ("man", complete_from_man),
